+++ src/modules/converter.js
@@ -1,0 +1,7 @@
+const toCelsius = inFahr => {
+  const inCels = inFahr - 273.15;
+  const roundDegree = Math.round(inCels);
+  return roundDegree;
+};
+
+export default toCelsius;
